Move SuggestionInputActivity from inner package to main package.
