--- sample/src/main/java/com/trendyol/uicomponents/suggestioninput/SuggestionInputViewState.kt
@@ -1,12 +1,0 @@
-package com.trendyol.uicomponents.suggestioninput
-
-import android.graphics.drawable.Drawable
-
-class SuggestionInputViewState(
-    val selectedBackground: Drawable?,
-    val unselectedBackground: Drawable?,
-    val selectedTextColor: Int,
-    val unselectedTextColor: Int,
-    val textSize: Float,
-    val padding: Float
-)
